[Refactor] moved changes from views/screensaver to gui/screens/screen. views/screensaver is removed.
